Add missing intptr_t cast in openNativeFileDescriptor return
openNativeFileDescriptor returned (jlong) info while every other JNI
open method goes through (jlong) (intptr_t) info. Pointer-to-wider-
integer conversion is implementation-defined in C; route it through
intptr_t for a guaranteed-correct round-trip and consistency with the
rest of the file.

Fixes #871

+++ android-gif-drawable/src/main/c/gif.c
@@ -354,7 +354,7 @@ Java_pl_droidsonroids_gif_GifInfoHandle_openNativeFileDescriptor(JNIEnv *env, jc
 		if (info == NULL) {
 			fclose(file);
 		}
-		return (jlong) info;
+		return (jlong) (intptr_t) info;
 	} else {
 		throwGifIOException(D_GIF_ERR_OPEN_FAILED, env, true);
 		close(fd);
